--- examples/demo/pyproject.toml
@@ -1,6 +1,0 @@
-[project]
-name = "logspyq_demo_agent"
-version = "0.1.0"
-
-[project.entry-points."logspyq.agents"]
-logspyq_demo_agent = "logspyq_demo_agent"
